+++ src/components/BaseComponents/BaseModal/index.vue
@@ -32,7 +32,6 @@
     export default {
         components: {
             BaseButton,
-
         }
     }
 </script>
