+++ src/components/BaseComponents/BaseTextarea/index.vue
@@ -40,8 +40,6 @@
                         max: 500,
                     }
                 }
-
-
             }
         },
         methods: {
