+++ src/router.js
@@ -6,25 +6,12 @@ import AddLot from './views/AddLot'
 import SuccessfulPublishment from './views/LotSuccessfulPublishment'
 import LotsPreview from './views/LotsPreview'
 
-Vue.use(Router)
+Vue.use(Router);
 
 export default new Router({
     mode: 'history',
     base: process.env.BASE_URL,
     routes: [
-        /* {
-             path: '/',
-             name: 'home',
-             component: Home
-         },
-       /*{
-             path: '/about',
-             name: 'about',
-             // route level code-splitting
-             // this generates a separate chunk (about.[hash].js) for this route
-             // which is lazy-loaded when the route is visited.
-             component: () => import(/!* webpackChunkName: "about" *!/ './views/About.vue')
-         },*/
         {
             path: '/auction',
             component: AuctionLots,
@@ -38,7 +25,6 @@ export default new Router({
                 {
                     path: '',
                     component: LotsPreview,
-                    // name: 'Lotdetailes',
                 },
 
             ]
